Extract events
