+++ packages/ciphernode/events/src/enclave_event/ciphertext_output_published.rs
@@ -1,0 +1,17 @@
+use crate::E3id;
+use actix::Message;
+use serde::{Deserialize, Serialize};
+use std::fmt::{self, Display};
+
+#[derive(Message, Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[rtype(result = "()")]
+pub struct CiphertextOutputPublished {
+    pub e3_id: E3id,
+    pub ciphertext_output: Vec<u8>,
+}
+
+impl Display for CiphertextOutputPublished {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "e3_id: {}", self.e3_id,)
+    }
+}
