+++ packages/ciphernode/events/src/enclave_event/test_event.rs
@@ -1,0 +1,17 @@
+use actix::Message;
+use serde::{Deserialize, Serialize};
+use std::fmt::{self, Display};
+
+#[derive(Message, Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[rtype(result = "()")]
+pub struct TestEvent {
+    pub msg: String,
+    pub entropy: u64,
+}
+
+#[cfg(test)]
+impl Display for TestEvent {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "TestEvent(msg: {})", self.msg)
+    }
+}
