+++ packages/ciphernode/events/src/lib.rs
@@ -1,9 +1,15 @@
+mod e3id;
+mod enclave_event;
+mod event_id;
 mod eventbus;
-mod events;
 mod ordered_set;
+mod seed;
 mod tag;
 
+pub use e3id::*;
+pub use enclave_event::*;
+pub use event_id::*;
 pub use eventbus::*;
-pub use events::*;
 pub use ordered_set::*;
+pub use seed::*;
 pub use tag::*;
